pyproject.toml
[project]
name = "code-review-agent"
version = "0.1.0"
description = "AI-powered code review agent that automatically reviews pull requests"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.68.0",
    "requests>=2.31.0",
    "PyGithub>=1.59.0",
    "python-dotenv>=1.0.0",
    "quart>=0.19.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]


[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
